消除警告，修复窗口最小化时窗口比例计算问题

+++ Yuicy/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -83,7 +83,7 @@ namespace Yuicy {
 					ShaderDataTypeToOpenGLBaseType(element.Type),
 					element.Normalized ? GL_TRUE : GL_FALSE,
 					layout.GetStride(),
-					(const void*)element.Offset);
+					(const void*)(uint32_t)element.Offset);
 				m_VertexBufferIndex++;
 				break;
 			}
